requirements.txt
fastapi 
uvicorn 
supabase
pydantic-settings
python-dotenv
psycopg2
torch
transformers
opencv-python
matplotlib
numpy
pillow
pandas
python-multipart